requirements.txt
# Core Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21
sqlite3

# AI/ML Libraries
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.7.2

# Data Handling
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Utilities
uuid
datetime
json
os
logging

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production Server
gunicorn==21.2.0